Refactor: extract chat item in its own component
