+++ src/features/messenger/ui/ChatItem.tsx
@@ -1,0 +1,37 @@
+import clsx from 'clsx'
+import { UserCard } from '@/shared/ui'
+import { MessageItem } from '../api'
+import { MeResponse } from '@/features/auth/api/authApi.types'
+import { modifyMessage } from '../utils'
+import s from './ChatItem.module.scss'
+
+type Props = {
+  chat: MessageItem
+  currentUser?: MeResponse
+  selectedId?: string
+  onSelect: (dialoguePartnerId: number) => void
+}
+
+export const ChatItem = ({ chat, currentUser, selectedId, onSelect }: Props) => {
+  const user = { id: chat.ownerId, userName: chat.userName, avatar: chat.avatars?.[0]?.url || '' }
+
+  const displayText = modifyMessage({
+    text: chat.messageText,
+    ownerId: chat.ownerId,
+    currentUserId: currentUser?.userId ?? -1,
+  })
+
+  const dialoguePartnerId = chat.ownerId === currentUser?.userId ? chat.receiverId : chat.ownerId
+
+  return (
+    <li
+      key={chat.id}
+      className={clsx(s.chatItem, selectedId === chat.receiverId.toString() && s.selected)}
+      onClick={() => onSelect(dialoguePartnerId)}
+    >
+      <UserCard layout="stacked" user={user} date={chat.createdAt}>
+        {displayText}
+      </UserCard>
+    </li>
+  )
+}
